--- obci/control/gui/presets/p300_MD.ini
@@ -1,30 +1,0 @@
-[P300 Calibration]
-info=...
-launch_file=scenarios/p300_MD/calibration_p300.ini
-public_params=
-category=P300
-
-[P300 Calibration - dummy]
-info=...
-launch_file=scenarios/p300_MD/calibration_p300_dummy.ini
-public_params=
-category=P300
-
-[P300 Offline Calibration]
-info=...
-launch_file=scenarios/p300_MD/calibration_p300_offline.ini
-public_params=
-category=P300
-
-[P300 Labyrinth]
-info=...
-launch_file=scenarios/p300_MD/p300_labyrinth.ini
-public_params=
-category=P300
-
-[P300 Labyrinth - dummy]
-info=...
-launch_file=scenarios/p300_MD/p300_labyrinth_dummy.ini
-public_params=
-category=P300
-
