+++ obci/interfaces/bci/p300_MD/helper_functions.py
@@ -5,12 +5,114 @@
 from scipy import *
 from scipy import linalg
 import numpy as np
+import pylab as pb
 from scipy import signal
+import scipy.stats
 import copy
 from obci.analysis.obci_signal_processing.signal import read_info_source
 from obci.analysis.obci_signal_processing.signal import read_data_source
 from obci.analysis.obci_signal_processing.tags import read_tags_source
 from obci.analysis.obci_signal_processing import read_manager
+from obci.analysis.obci_signal_processing.smart_tags_manager import SmartTagsManager
+from obci.analysis.obci_signal_processing.tags.smart_tag_definition import SmartTagDurationDefinition
+
+def target_tags_func(tag):
+    return tag['desc'][u'index']==tag['desc'][u'target']
+
+def nontarget_tags_func(tag):
+    return tag['desc'][u'index']!=tag['desc'][u'target']
+
+def get_epochs_fromfile(ds, start_offset=-0.1,duration=2.0, 
+                        filter=None, montage=None,
+                        drop_chnls = [ u'AmpSaw', u'DriverSaw', u'trig1', u'trig2']):
+    '''For offline calibration and testing, load target and nontarget
+    epochs using obci read_manager.
+    Args:
+        ds: dataset file name without extension.
+        start_offset: baseline in negative seconds,
+        duration: duration of the epoch (including baseline),
+        filter: list of [wp, ws, gpass, gstop] for scipy.signal.iirdesign
+                in Hz, Db, or None if no filtering is required
+        montage: list of ['montage name', ...] ...-channel names if required
+            montage name can be 'ears', 'csa', 'custom'
+            ears require 2 channel names for ear channels
+            custom requires list of reference channel names
+    Return: two lists of smart tags: target_tags, nontarget_tags'''
+    eeg_rm = read_manager.ReadManager(ds+'.xml', ds+'.raw', ds+'.tag')
+    eeg_rm = exclude_channels(eeg_rm, drop_chnls)
+
+    if filter:
+        eeg_rm = mgr_filter(eeg_rm, filter[0], filter[1],filter[2], 
+                            filter[3], ftype='cheby2', use_filtfilt=True)
+    if montage:
+        if montage[0] == 'ears':
+            eeg_rm = montage_ears(eeg_rm, montage[1], montage[2])
+        elif montage[0] == 'csa':
+            eeg_rm = montage_csa(eeg_rm)
+        elif montage[0] == 'custom':
+            eeg_rm = montage_custom(eeg_rm, montage[1:])
+        else:
+            raise Exception('Unknown montage')
+   
+    
+    tag_def = SmartTagDurationDefinition(start_tag_name=u'blink',
+                                        start_offset=start_offset,
+                                        end_offset=0.0,
+                                        duration=duration)
+    stags = SmartTagsManager(tag_def, '', '' ,'', p_read_manager=eeg_rm)
+    target_tags = stags.get_smart_tags(p_func = target_tags_func, p_from = 60.0, p_len=21400.0*512)
+    nontarget_tags = stags.get_smart_tags(p_func = nontarget_tags_func, p_from = 60.0, p_len=21400.0*512)
+    
+    return target_tags, nontarget_tags
+    
+def evoked_from_smart_tags(tags, chnames, bas = -0.1):
+    '''
+    Args:
+        tags: smart tag list, to average
+        chnames: list of channels to use for averaging,
+        bas: baseline (in negative seconds)'''
+    min_length = min(i.get_samples().shape[1] for i in tags)
+    # really don't like this, but epochs generated by smart tags can vary in length by 1 sample
+    channels_data = [] 
+    Fs = float(tags[0].get_param('sampling_frequency'))
+    for i in tags:
+        data = i.get_channels_samples(chnames)[:,:min_length]
+        for nr, chnl in enumerate(data):
+            data[nr] = chnl - np.mean(chnl[0:-Fs*bas])# baseline correction
+        if np.max(np.abs(data))<4000:
+            channels_data.append(data)
+    
+    return np.mean(channels_data, axis=0), scipy.stats.sem(channels_data, axis=0)
+    
+    
+def evoked_pair_plot_smart_tags(tags1, tags2, chnames=['O1', 'O2', 'Pz', 'PO7', 'PO8', 'PO3', 'PO4', 'Cz',],
+                                start_offset=-0.1, labels=['target', 'nontarget']):
+    '''debug evoked potential plot,
+     pairwise comparison of 2 smarttag lists,
+     blocks thread
+     Args:
+        chnames: channels to plot
+        start_offset: baseline in seconds'''
+    ev1, std1 = evoked_from_smart_tags(tags1, chnames, start_offset)
+    ev2, std2 = evoked_from_smart_tags(tags2, chnames, start_offset)
+    Fs = float(tags1[0].get_param('sampling_frequency'))
+    time = np.linspace(0+start_offset, ev1.shape[1]/Fs+start_offset, ev1.shape[1])
+    pb.figure()
+    for nr, i in enumerate(chnames):
+        pb.subplot( (len(chnames)+1)/2, 2, nr+1)
+        pb.plot(time, ev1[nr], 'r',label = labels[0]+' N:{}'.format(len(tags1)))
+        pb.fill_between(time, ev1[nr]-std1[nr], ev1[nr]+std1[nr],
+                            color = 'red', alpha=0.3, )
+        pb.plot(time, ev2[nr], 'b', label = labels[1]+' N:{}'.format(len(tags2)))
+        pb.fill_between(time, ev2[nr]-std2[nr], ev2[nr]+std2[nr],
+                        color = 'blue', alpha=0.3)
+        
+        pb.title(i)
+    pb.legend()
+    
+    pb.show()
+
+
 
 def mgr_filter(mgr, wp, ws, gpass, gstop, analog=0, ftype='ellip', output='ba', unit='hz', use_filtfilt=False, meancorr=1.0):
     if unit == 'radians':
