--- obci/interfaces/bci/p300_MD/p300_offline_learning_peer.ini
@@ -1,7 +1,0 @@
-[local_params]
-data_file_name=test1
-data_file_path=~
-montage=custom
-montage_channels=Cz
-offline = 1
-cl_filename = class
