--- obci/interfaces/bci/p300_MD/p300_online_decision_peer.ini
@@ -1,17 +1,0 @@
-[local_params]
-montage=custom
-montage_channels=Cz
-cl_filename = class
-hold_after_dec = 1
-data_file_path = ~
-data_file_name=test1
-cl_filename = class
-
-[config_sources]
-logic=
-amplifier=
-
-[external_params]
-blink_field_ids=logic.active_field_ids
-sampling_rate=amplifier.sampling_rate
-channel_names=amplifier.channel_names
