+++ obci/scenarios/budzik/prototypes/p300_configs/p300_labyrinth_signal_saver_config.ini
@@ -1,0 +1,2 @@
+[local_params]
+save_file_name=test2
