+++ obci/scenarios/budzik/prototypes/p300_configs/p300_labyrinth_switch_config.ini
@@ -1,0 +1,2 @@
+[local_params]
+finish_saving=1
