+++ obci/scenarios/budzik/prototypes/p300_labyrinth_dummy.ini
@@ -16,7 +16,7 @@ config=scenarios/budzik/prototypes/p300_configs/cap_brain2013_dummy.ini
 ;***********************************************
 [peers.signal_saver]
 path=acquisition/signal_saver_peer.py
-config=scenarios/p300_MD/configs/p300_labyrinth_signal_saver_config.ini
+config=scenarios/budzik/prototypes/p300_configs/p300_labyrinth_signal_saver_config.ini
 
 [peers.signal_saver.launch_dependencies]
 amplifier=amplifier
@@ -86,7 +86,7 @@ analysis=analysis
 ;***********************************************
 [peers.switch_backup]
 path=interfaces/switch/backup/switch_backup_peer.py
-config=scenarios/p300_MD/configs/p300_labyrinth_switch_config.ini
+config=scenarios/budzik/prototypes/p300_configs/p300_labyrinth_switch_config.ini
 
 ;***********************************************
 [peers.switch]
