--- obci/scenarios/p300_MD/calibration_p300.ini
@@ -1,69 +1,0 @@
-[peers]
-scenario_dir=
-;***********************************************
-[peers.mx]
-path=multiplexer-install/bin/mxcontrol
-
-;***********************************************
-[peers.config_server]
-path=control/peer/config_server.py
-
-;***********************************************
-[peers.amplifier]
-path=drivers/eeg/cpp_amplifiers/amplifier_tmsi.py
-config=scenarios/brain2013/configs/cap_brain2013.ini
-
-;***********************************************
-[peers.signal_saver]
-path=acquisition/signal_saver_peer.py
-
-[peers.signal_saver.launch_dependencies]
-amplifier=amplifier
-
-;***********************************************
-[peers.info_saver]
-path=acquisition/info_saver_peer.py
-
-[peers.info_saver.launch_dependencies]
-amplifier=amplifier
-signal_saver=signal_saver
-
-;***********************************************
-[peers.tag_saver]
-path=acquisition/tag_saver_peer.py
-
-[peers.tag_saver.launch_dependencies]
-signal_saver=signal_saver
-
-
-;***********************************************
-[peers.ugm_engine]
-path=gui/ugm/blinking/ugm_blinking_engine_peer.py
-config=scenarios/brain2013/configs/p300_fast_colour_calibration.ini
-
-;***********************************************
-[peers.ugm_server]
-path=gui/ugm/ugm_server_peer.py
-
-[peers.ugm_server.launch_dependencies]
-ugm_engine=ugm_engine
-
-;***********************************************
-[peers.logic]
-path=interfaces/bci/p300_fda/logic_p300_calibration_peer.py
-config=scenarios/brain2013/configs/brain2013_logic_p300_calibraction.ini
-
-[peers.logic.launch_dependencies]
-ugm_engine=ugm_engine
-ugm_server=ugm_server
-signal_saver=signal_saver
-
-;***********************************************
-[peers.clasifier]
-path=interfaces/bci/p300_MD/p300_offline_learning_peer.py
-config=scenarios/p300_MD/configs/calibration_p300_clasifier_config.ini
-
-[peers.clasifier.launch_dependencies]
-signal_saver=signal_saver
-logic=logic
-
