--- obci/scenarios/p300_MD/calibration_p300_offline.ini
@@ -1,21 +1,0 @@
-
-
-[peers]
-scenario_dir=
-;***********************************************
-[peers.mx]
-path=multiplexer-install/bin/mxcontrol
-
-;***********************************************
-[peers.config_server]
-path=control/peer/config_server.py
-
-;***********************************************
-
-[peers.clasifier]
-path=interfaces/bci/p300_MD/p300_offline_learning_peer.py
-config=scenarios/p300_MD/configs/offline_calibration_p300_clasifier_config.ini
-
-
-
-
