--- obci/scenarios/p300_MD/configs/calibration_p300_clasifier_config.ini
@@ -1,19 +1,0 @@
-[local_params]
-
-montage=custom
-montage_channels=Cz
-offline = 0
-cl_filename = test1_class.dump
-
-[config_sources]
-signal_saver=
-logic=
-
-[external_params]
-
-data_file_name=signal_saver.save_file_name
-data_file_path=signal_saver.save_file_path
-
-[launch_dependencies]
-signal_saver=
-logic=
