--- obci/scenarios/p300_MD/configs/offline_calibration_p300_clasifier_config.ini
@@ -1,13 +1,0 @@
-[local_params]
-
-montage=custom
-montage_channels=Cz
-offline = 1
-data_file_name=test1
-data_file_path=~
-cl_filename = test1_class.dump
-
-
-
-
-
