--- obci/scenarios/p300_MD/configs/p300_labyrinth_clasifier_config.ini
@@ -1,10 +1,0 @@
-[local_params]
-data_file_name=test1
-data_file_path=~
-cl_filename = test1_class.dump
-
-[config_sources]
-logic=
-
-[external_params]
-blink_field_ids=logic.active_field_ids
