--- obci/scenarios/p300_MD/configs/p300_labyrinth_signal_saver_config.ini
@@ -1,2 +1,0 @@
-[local_params]
-save_file_name=test2
