--- obci/scenarios/p300_MD/configs/p300_labyrinth_switch_config.ini
@@ -1,2 +1,0 @@
-[local_params]
-finish_saving=1
