--- obci/scenarios/p300_MD/p300_labyrinth.ini
@@ -1,96 +1,0 @@
-[peers]
-scenario_dir=
-;***********************************************
-[peers.mx]
-path=multiplexer-install/bin/mxcontrol
-
-;***********************************************
-[peers.config_server]
-path=control/peer/config_server.py
-
-;***********************************************
-[peers.amplifier]
-path=drivers/eeg/cpp_amplifiers/amplifier_tmsi.py
-config=scenarios/brain2013/configs/cap_brain2013.ini
-
-;***********************************************
-[peers.signal_saver]
-path=acquisition/signal_saver_peer.py
-config=scenarios/p300_MD/configs/p300_labyrinth_signal_saver_config.ini
-
-[peers.signal_saver.launch_dependencies]
-amplifier=amplifier
-
-;***********************************************
-[peers.tag_saver]
-path=acquisition/tag_saver_peer.py
-
-[peers.tag_saver.launch_dependencies]
-signal_saver=signal_saver
-
-;***********************************************
-[peers.info_saver]
-path=acquisition/info_saver_peer.py
-
-[peers.info_saver.launch_dependencies]
-amplifier=amplifier
-signal_saver=signal_saver
-
-;***********************************************
-[peers.ugm_engine]
-path=gui/ugm/blinking/ugm_blinking_engine_peer.py
-config=scenarios/brain2013/configs/p300_fast_colour_bci.ini
-
-[peers.ugm_engine.config_sources]
-logic=logic
-
-;***********************************************
-[peers.ugm_server]
-path=gui/ugm/ugm_server_peer.py
-
-[peers.ugm_server.launch_dependencies]
-ugm_engine=ugm_engine
-
-;***********************************************
-[peers.analysis]
-path=interfaces/bci/p300_MD/p300_online_decision_peer.py
-config=scenarios/p300_MD/configs/p300_labyrinth_clasifier_config.ini
-
-[peers.analysis.config_sources]
-logic=logic
-amplifier=amplifier
-
-[peers.analysis.launch_dependencies]
-logic=logic
-amplifier=amplifier
-
-;*************************************************
-[peers.logic]
-path=logic/logic_maze_peer.py
-config=scenarios/brain2013/configs/brain2013_logic_maze_peer.ini
-
-[peers.logic.launch_dependencies]
-ugm=ugm_server
-signal_saver=signal_saver
-
-;************************************
-[peers.feedback]
-path=logic/feedback/logic_decision_feedback_peer.py
-
-[peers.feedback.launch_dependencies]
-ugm_engine=ugm_engine
-ugm_server=ugm_server
-logic=logic
-analysis=analysis
-
-;***********************************************
-[peers.switch_backup]
-path=interfaces/switch/backup/switch_backup_peer.py
-config=scenarios/p300_MD/configs/p300_labyrinth_switch_config.ini
-
-;***********************************************
-[peers.switch]
-path=drivers/switch/switch_amplifier_peer.py
-
-[peers.switch.launch_dependencies]
-ugm_engine=ugm_engine
